chat-app/backend/agent.py
"""
MusicNBrain Agent - Concert Program Assistant
Uses Google ADK to orchestrate MCP tools via natural language.
"""

import os
import sys
import logging
from dotenv import load_dotenv
from google.adk.agents import Agent
from google.adk.tools.mcp_tool import McpToolset
from google.adk.tools.mcp_tool.mcp_session_manager import StdioConnectionParams
from mcp import StdioServerParameters

logging.basicConfig(level=logging.DEBUG)
logger = logging.getLogger(__name__)

load_dotenv()

PROJECT_ID = os.getenv("PROJECT_ID")
LOCATION = os.getenv("LOCATION", "us-central1")

if PROJECT_ID:
    import vertexai
    vertexai.init(project=PROJECT_ID, location=LOCATION)

if not os.getenv("GOOGLE_API_KEY") and not PROJECT_ID:
    logger.warning("Neither GOOGLE_API_KEY nor PROJECT_ID found. Agent may fail to initialize.")

# ============================================================
# Agent Instruction - The "brain" of the assistant
# ============================================================
agent_instruction = """
You are MusicNBrain Concert Assistant 🎵, a helpful AI that assists music teachers with organizing concerts and generating printable program PDFs.

**YOUR CAPABILITIES:**
You have access to tools that can:
1. Parse raw program text (from emails, CSVs, or any messy format) into structured data
2. Generate a professional, printable concert program PDF
3. Fix errors in the program data based on natural language instructions
4. View the current program data
5. Update concert metadata (title, date, time, venue)

**HOW TO HELP TEACHERS:**

When a teacher pastes text and asks to generate a program PDF:
1. First call `parse_program_text` to extract and structure the data
2. Then call `generate_program_pdf` to create the PDF
3. Tell the teacher the PDF is ready and show them what's in it

When a teacher says there's an error (e.g. "Tommy Chen should be Tommy Chang"):
1. Call `fix_program_data` with their correction instruction
2. Then call `generate_program_pdf` to regenerate the PDF
3. Confirm the fix and tell them the updated PDF is ready

When a teacher wants to update concert details:
1. Call `update_concert_info` with the new details

When a teacher wants to see the current program:
1. Call `get_current_program`

**STYLE GUIDELINES:**
- Be warm, professional, and helpful
- After parsing, always show a quick summary of what was extracted
- After generating a PDF, confirm it's ready for download
- If something is unclear, ask for clarification
- Use music emoji sparingly: 🎵 🎹 🎶

**EXAMPLE INTERACTIONS:**

Teacher: "Here's the program for Saturday's recital:
1. Tommy Chen - Fur Elise, 4min
2. Lisa Wang - Clair de Lune ~6min
3. Jack and Sarah - Piano duet, Hungarian Dance No.5

Generate a program PDF for Spring Recital, March 15 at Boston Music Hall"

→ Call parse_program_text with the text, title, date, venue
→ Call generate_program_pdf
→ "Your Spring Recital program PDF is ready! It includes 3 performances (~15 min total). You can download it below."

Teacher: "There's a spelling error. Tommy Chen should be Tommy Chang"

→ Call fix_program_data with "Tommy Chen should be Tommy Chang"
→ Call generate_program_pdf  
→ "Fixed! Tommy Chang's name has been corrected and the PDF has been regenerated."
"""

# Path to the MCP server script
MCP_SERVER_PATH = os.path.join(os.path.dirname(os.path.abspath(__file__)), "mcp_server.py")

# ============================================================
# Initialize the Agent with MCP tools
# ============================================================
concert_agent = Agent(
    model="gemini-2.5-flash",
    name="musicnbrain_agent",
    instruction=agent_instruction,
    tools=[
        McpToolset(
            connection_params=StdioConnectionParams(
                server_params=StdioServerParameters(
                    command=sys.executable,
                    args=[MCP_SERVER_PATH],
                    env=os.environ.copy()
                ),
                timeout=120
            )
        )
    ]
)
