chat-app/backend/pyproject.toml
[project]
name = "musicnbrain-backend"
version = "0.1.0"
description = "MusicNBrain Concert Assistant Backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.122.0",
    "google-adk>=1.18.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
    "mcp>=1.1.2",
    "google-genai",
    "fastmcp",
    "reportlab>=4.0",
]

[[tool.uv.index]]
url = "https://pypi.org/simple"
default = true
